% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShippyOrders.R
\name{getShippyOrders}
\alias{getShippyOrders}
\title{ShipStation Orders Endpoint}
\usage{
getShippyOrders(shippyPath, apiKey, apiSecret, verbose = T, page = NULL)
}
\arguments{
\item{shippyPath}{Something like: https://api.mywebsite.com/}

\item{apiKey}{Unencoded api key}

\item{verbose}{Whether it will return the results of the api call. Defaults to T.}

\item{page}{Returns a specific page of results. Defaults to 1.}

\item{apiPassword}{Unencoded apiPassword}
}
\value{
Returns a list with (1) the number of pages, and (2) a dataframe of orders.
}
\description{
Gets shipStation Orders

Documentation: # https://www.shipstation.com/developer-api/#/reference/customers/list-customers/list-customers
}
